crypto/cryptodev: expansion hash operation
(1) remove size restriction for single hash operation
(2) support hash operation to update uint32_t data

+++ include/crypto/xform.h
@@ -59,7 +59,7 @@ struct auth_hash
   CODE void (*init) (FAR void *);
   CODE void (*setkey) (FAR void *, FAR const uint8_t *, uint16_t);
   CODE void (*reinit) (FAR void *, FAR const uint8_t *, uint16_t);
-  CODE int  (*update) (FAR void *, FAR const uint8_t *, uint16_t);
+  CODE int  (*update) (FAR void *, FAR const uint8_t *, size_t);
   CODE void (*final) (FAR uint8_t *, FAR void *);
 };
